Fixed two mistakes in the README.md file
The example application should be opened in two or three browser windows (not in "several"). Handling the video rendering and data transfer of more OakStreaming instances simultaneously seem to be quite challenging for most notebooks (Probably not due to bad programming of the OakStreaming client).
Moreover, fixed one spelling mistake.